% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{calc_bbox}
\alias{calc_bbox}
\title{Calculate the bounding box for the observation}
\usage{
calc_bbox(location, f = 0.25)
}
\arguments{
\item{location}{the output from \code{parse_coords}}

\item{f}{how much space to include around the GPS coordinate}
}
\value{
a vector of length 4 that includes the coordinates of the corners of
the bounding box
}
\description{
Calculate the bounding box for the observation
}
