% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{draw_obs_map}
\alias{draw_obs_map}
\title{Draw the observation map}
\usage{
draw_obs_map(obs)
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}
}
\value{
The path of the file that includes the map as a png file
}
\description{
Draw the observation map
}
