% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_data.R
\name{get_inat_obs}
\alias{get_inat_obs}
\title{Transform the list coming back from the API into a tibble}
\usage{
get_inat_obs(obs)
}
\arguments{
\item{obs}{A data frame with the iNaturalist information}
}
\value{
a tibble
}
\description{
The tibble contains the information needed to compose the toot
}
