% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_data.R
\name{get_inat_obs_raw}
\alias{get_inat_obs_raw}
\title{Get raw data from the iNaturalist API}
\usage{
get_inat_obs_raw(n_res = 50, observed_on = as.character(Sys.Date() - 365))
}
\arguments{
\item{n_res}{number of results to return}

\item{observed_on}{date of when the observation was made (by default 1 year
before today's date)}
}
\value{
a list representing the JSON coming back from the API
}
\description{
This only returns observations that:
\itemize{
\item have photo(s) licensed cc-by or cc0
\item can be mapped
\item that are "research" quality grade
}
}
