% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_toot_text.R
\name{get_obs_emoji}
\alias{get_obs_emoji}
\title{Get the emoji that represents the taxa in the observation}
\usage{
get_obs_emoji(obs)
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}
}
\description{
Get the emoji that represents the taxa in the observation
}
