% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_toot_text.R
\name{get_obs_hashtags}
\alias{get_obs_hashtags}
\title{Get the hashtags for the iNaturalist observation}
\usage{
get_obs_hashtags(obs)
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}
}
\description{
Get the hashtags for the iNaturalist observation
}
