% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_photo.R
\name{get_obs_photo}
\alias{get_obs_photo}
\title{Get the photo associated associated with the observation}
\usage{
get_obs_photo(obs)
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}
}
\value{
The path of the image downloaded from iNaturalist
}
\description{
Get the photo associated associated with the observation
}
