% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_map.R
\name{parse_coords}
\alias{parse_coords}
\title{Convert GPS coordinates from iNat into latitude and longitude}
\usage{
parse_coords(location)
}
\arguments{
\item{location}{the GPS coordinate string from iNat (in the form "lat,lon")}
}
\description{
Convert GPS coordinates from iNat into latitude and longitude
}
