% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_obs.R
\name{pick_obs}
\alias{pick_obs}
\title{Pick observations to post}
\usage{
pick_obs(obs, n_obs)
}
\arguments{
\item{obs}{data frame returned by \code{get_inat_obs}}

\item{n_obs}{number of indices to return}
}
\value{
a vector of indices corresponding to the number of observations to
return
}
\description{
Pick observations to post
}
