% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_natbot_toot.R
\name{post_natbot_toot}
\alias{post_natbot_toot}
\title{Write the Toots with the iNaturalist information}
\usage{
post_natbot_toot(obs, scheduled_at = NULL, token = "RTOOT_TOKEN")
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}

\item{scheduled_at}{Specify date/time for posting the toot}

\item{token}{the name of the environment variable that holds the Mastodon
token}
}
\description{
Write the Toots with the iNaturalist information
}
