% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inat_data.R
\name{write_obs_data}
\alias{write_obs_data}
\title{Write iNaturalist Observation data to CSV file}
\usage{
write_obs_data(obs, path = ".")
}
\arguments{
\item{obs}{The data frame returned by \code{get_inat_obs}}

\item{path}{The path where the CSV file should be written}
}
\value{
the full path of the CSV file
}
\description{
Write iNaturalist Observation data to CSV file
}
