% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_alt_text.R
\name{write_photo_alt_text}
\alias{write_photo_alt_text}
\title{Write the alt text accompanying the photo observation}
\usage{
write_photo_alt_text(obs)
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}
}
\description{
Write the alt text accompanying the photo observation
}
