% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_toot_text.R
\name{write_toot_text}
\alias{write_toot_text}
\title{Compose the content of the toot based on the information associated with the observation}
\usage{
write_toot_text(obs)
}
\arguments{
\item{obs}{A one-row data frame with the iNaturalist information}
}
\value{
A string with the text for the toot
}
\description{
Compose the content of the toot based on the information associated with the observation
}
